dock/distributed-task-system/worker/requirements.txt
redis
psycopg2-binary
